Re-organized code files
