+++ client/webapp/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "eslint": "8.28.0",
     "eslint-config-next": "13.0.4",
+    "js-cookie": "^3.0.1",
     "next": "13.0.4",
     "react": "18.2.0",
     "react-dom": "18.2.0"
