+++ client/webapp/pages/api/login.js
@@ -1,4 +1,4 @@
-import { sendServerRequest } from "../../utils/serverUtils";
+import { sendServerRequest } from "../../src/utils/serverUtils";
 
 export default async function handler(req, res) {
   if (req.method !== "POST")
