--- client/webapp/utils/imageUtils.js
@@ -1,3 +1,0 @@
-export function unsplashLoader({ src, width, quality }) {
-  return src;
-}
